handler: host landing page under `/about/`
serve the landing page from a dedicated route under, which does
not collide with the authentication handler. this allows us
to leave the HTTP method validation to the ServerMux.

the root path (and /index.html for good measure) are redirecting
to the new location to provide a convenient way to access the
information. the root path is registered with `{$}` to prevent
ServerMux from treating the handler as catch-all.