mock: refactor storage system to support custom token prefixes
replace the static DTO storage with a data access implementation
which uses memory collections for storage. proper interfaces allow
future expansions into real SQL territory.

this change removes support for "lenient" token matching; all auth
requests must match an existing token as-is. this resembles the
reference API better. the feature was initially implemented for
convenience, but since we already have ready-to-use payloads in
testdata, there really no more use for this feature.

the number of "feature" groups can also be adjusted via commandline
arguments, to allow testing scenarios where users are in more groups
than the api is willing to return in a single page.