auth: implement token preflight validation
prevent unecessary calls to the Gitlab API if the token is obviously
not issued by the same. since Gitlab allows customizing the prefix
for issued tokens, the service can be configured to accept different
kind of token prefixes. mismatches are rejected early without contacting
the Gitlab API at all. this should reduce traffic and therefor ratelimiting
issues when calling the API with too many invalid tokens.